separate spin and particle numbertests
